refactor : ExceptionType 을 외부 API 의 오류와 구분하기 위해 RoomescapeException 으로 변경
