+++ src/main/java/roomescape/domain/Password.java
@@ -1,6 +1,6 @@
 package roomescape.domain;
 
-import static roomescape.exception.ExceptionType.INVALID_PASSWORD;
+import static roomescape.exception.RoomescapeExceptionType.INVALID_PASSWORD;
 
 import jakarta.persistence.Embeddable;
 import java.util.Objects;
