+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,9 +1,9 @@
 package roomescape.domain;
 
-import static roomescape.exception.ExceptionType.EMPTY_DATE;
-import static roomescape.exception.ExceptionType.EMPTY_MEMBER;
-import static roomescape.exception.ExceptionType.EMPTY_THEME;
-import static roomescape.exception.ExceptionType.EMPTY_TIME;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_DATE;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_MEMBER;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_THEME;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_TIME;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
