+++ src/main/java/roomescape/domain/Role.java
@@ -1,7 +1,7 @@
 package roomescape.domain;
 
 import java.util.Arrays;
-import roomescape.exception.ExceptionType;
+import roomescape.exception.RoomescapeExceptionType;
 import roomescape.exception.RoomescapeException;
 
 public enum Role {
@@ -18,7 +18,7 @@ public enum Role {
         return Arrays.stream(values())
                 .filter(role -> role.tokenValue.equals(value))
                 .findFirst()
-                .orElseThrow(() -> new RoomescapeException(ExceptionType.NOT_FOUND_ROLE));
+                .orElseThrow(() -> new RoomescapeException(RoomescapeExceptionType.NOT_FOUND_ROLE));
     }
 
     public boolean isAdmin() {
