+++ src/main/java/roomescape/service/ThemeService.java
@@ -1,7 +1,7 @@
 package roomescape.service;
 
-import static roomescape.exception.ExceptionType.DELETE_USED_THEME;
-import static roomescape.exception.ExceptionType.DUPLICATE_THEME;
+import static roomescape.exception.RoomescapeExceptionType.DELETE_USED_THEME;
+import static roomescape.exception.RoomescapeExceptionType.DUPLICATE_THEME;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
