+++ src/test/java/roomescape/domain/EmailTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.exception.ExceptionType;
+import roomescape.exception.RoomescapeExceptionType;
 import roomescape.exception.RoomescapeException;
 
 class EmailTest {
@@ -22,7 +22,7 @@ class EmailTest {
     void notEmailPatternExceptionTest(String wrongEmail) {
         assertThatThrownBy(() -> new Email(wrongEmail))
                 .isInstanceOf(RoomescapeException.class)
-                .hasMessage(ExceptionType.INVALID_EMAIL.getMessage());
+                .hasMessage(RoomescapeExceptionType.INVALID_EMAIL.getMessage());
     }
 
     @DisplayName("올바른 이메일 형식으로 Email 객체를 생성할 수 있다.")
