+++ src/test/java/roomescape/domain/NameTest.java
@@ -2,7 +2,7 @@ package roomescape.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static roomescape.exception.ExceptionType.INVALID_NAME;
+import static roomescape.exception.RoomescapeExceptionType.INVALID_NAME;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
