+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -3,10 +3,10 @@ package roomescape.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static roomescape.exception.ExceptionType.EMPTY_DATE;
-import static roomescape.exception.ExceptionType.EMPTY_MEMBER;
-import static roomescape.exception.ExceptionType.EMPTY_THEME;
-import static roomescape.exception.ExceptionType.EMPTY_TIME;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_DATE;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_MEMBER;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_THEME;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_TIME;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
