+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -3,7 +3,7 @@ package roomescape.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static roomescape.exception.ExceptionType.EMPTY_TIME;
+import static roomescape.exception.RoomescapeExceptionType.EMPTY_TIME;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
