+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static roomescape.exception.ExceptionType.DELETE_USED_TIME;
-import static roomescape.exception.ExceptionType.DUPLICATE_RESERVATION_TIME;
+import static roomescape.exception.RoomescapeExceptionType.DELETE_USED_TIME;
+import static roomescape.exception.RoomescapeExceptionType.DUPLICATE_RESERVATION_TIME;
 
 import java.time.LocalDate;
 import java.util.List;
